backend/requirements.txt
fastapi
uvicorn
openai
numpy
python-dotenv
paddleocr
Pillow
